gfx/layers/apz/test/mochitest/test_group_wheelevents.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Various wheel-scrolling tests that spawn in new windows</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="apz_test_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

SimpleTest.waitForExplicitFinish();

var subtests = [
  {'file': 'helper_scroll_on_position_fixed.html', 'prefs': prefs},
  {'file': 'helper_bug1271432.html', 'prefs': prefs},
  {'file': 'helper_scroll_inactive_perspective.html', 'prefs': prefs},
  {'file': 'helper_scroll_inactive_zindex.html', 'prefs': prefs}
];

if (isApzEnabled()) {
  SimpleTest.waitForExplicitFinish();
  window.onload = function() {
    runSubtestsSeriallyInFreshWindows(subtests)
    .then(SimpleTest.finish);
  };
}

window.onload = function() {
  if (!SpecialPowers.getDOMWindowUtils(window).asyncPanZoomEnabled) {
    ok(true, "APZ is not enabled, this test is not relevant, sorry!\n");
    SimpleTest.finish();
    return;
  }
  testDone();
};

  </script>
</head>
<body>
</body>
</html>
